+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix constructors not getting static methods when in non-native-modifying mode
 * Fix: `unmethodize` and `methodize` shouldn't get stuck in a loop because of the function name
 * Test: creating an anonymous function is a bit harder in new node versions
+* Improve `Object.alike` performance and add tests
 
 ## 0.3.3 (2016-10-14)